+++ src/BaseDto.php
@@ -6,8 +6,10 @@ namespace Crescat\SaloonSdkGenerator;
 
 use Crescat\SaloonSdkGenerator\Contracts\Deserializable;
 use Crescat\SaloonSdkGenerator\Traits\Deserializes;
+use Crescat\SaloonSdkGenerator\Traits\HasArrayableAttributes;
 
 abstract class BaseDto implements Deserializable
 {
     use Deserializes;
+    use HasArrayableAttributes;
 }
